new extension updates: optimizing

+++ extensions/dev.efaz.bundled_extension/resources/create_bundle.py
@@ -4,28 +4,17 @@ try:
     import subprocess
     import os
 except Exception as e:
-    print(
-        f"\033[38;5;196m{'Something went wrong trying to use provided python modules!'}\033[0m"
-    )
+    print(f"\033[38;5;196m{'Something went wrong trying to use provided python modules!'}\033[0m")
 
-def printMainMessage(mes):
-    print(f"\033[38;5;255m{mes}\033[0m")
-def printErrorMessage(mes):
-    print(f"\033[38;5;196m{mes}\033[0m")
-def printSuccessMessage(mes):
-    print(f"\033[38;5;82m{mes}\033[0m")
-def printWarnMessage(mes):
-    print(f"\033[38;5;202m{mes}\033[0m")
-def printYellowMessage(mes):
-    print(f"\033[38;5;226m{mes}\033[0m")
-def printDebugMessage(mes):
-    print(f"\033[38;5;226m{mes}\033[0m")
-def isYes(text):
-    return text.lower() == "y" or text.lower() == "yes"
-def isNo(text):
-    return text.lower() == "n" or text.lower() == "no"
-def isRequestClose(text):
-    return text.lower() == "exit" or text.lower() == "exit()"
+def printMainMessage(mes): print(f"\033[38;5;255m{mes}\033[0m")
+def printErrorMessage(mes): print(f"\033[38;5;196m{mes}\033[0m")
+def printSuccessMessage(mes): print(f"\033[38;5;82m{mes}\033[0m")
+def printWarnMessage(mes): print(f"\033[38;5;202m{mes}\033[0m")
+def printYellowMessage(mes): print(f"\033[38;5;226m{mes}\033[0m")
+def printDebugMessage(mes): print(f"\033[38;5;226m{mes}\033[0m")
+def isYes(text): return text.lower() == "y" or text.lower() == "yes"
+def isNo(text): return text.lower() == "n" or text.lower() == "no"
+def isRequestClose(text): return text.lower() == "exit" or text.lower() == "exit()"
 
 # Preparing Generation
 printWarnMessage("--- Preparing Generation ---")
@@ -60,19 +49,15 @@ printMainMessage(f"Extensions: {json.dumps(extensions_added)}")
 # Updating Bundle Information
 printWarnMessage("--- Updating Bundle Information ---")
 printMainMessage("Opening Settings JSON..")
-with open(os.path.join(extension_path, "settings.json"), "r") as f:
-    se = json.load(f)
+with open(os.path.join(extension_path, "settings.json"), "r") as f: se = json.load(f)
 printMainMessage("Opening Manifest JSON..")
-with open(os.path.join(extension_path, "manifest.json"), "r") as f:
-    ma = json.load(f)
+with open(os.path.join(extension_path, "manifest.json"), "r") as f: ma = json.load(f)
 mans = {}
 ses = {}
 for i in extensions_added:
     printMainMessage(f"Opening JSONs for {i}..")
-    with open(f"{extension_path}/{i}/manifest.json", "r") as f:
-        mans[i] = json.load(f)
-    with open(f"{extension_path}/{i}/settings.json", "r") as f:
-        ses[i] = json.load(f)
+    with open(f"{extension_path}/{i}/manifest.json", "r") as f: mans[i] = json.load(f)
+    with open(f"{extension_path}/{i}/settings.json", "r") as f: ses[i] = json.load(f)
 content_scripts = []
 content_css = []
 web_accessible_resources = []
@@ -85,15 +70,11 @@ for a, v in mans.items():
             jss = []
             csss = []
             for e in js:
-                if e.find(f"{a}/") == -1:
-                    jss.append(f"{a}/{e}")
-                else:
-                    jss.append(f"{e}")
+                if e.find(f"{a}/") == -1: jss.append(f"{a}/{e}")
+                else: jss.append(f"{e}")
             for e in css:
-                if e.find(f"{a}/") == -1:
-                    csss.append(f"{a}/{e}")
-                else:
-                    csss.append(f"{e}")
+                if e.find(f"{a}/") == -1: csss.append(f"{a}/{e}")
+                else: csss.append(f"{e}")
             content_scripts = content_scripts + jss
             content_css = content_css + csss
 for a, v in mans.items():
@@ -102,18 +83,14 @@ for a, v in mans.items():
             resou = i.get("resources", [])
             resourcess = []
             for e in resou:
-                if e.find(f"{a}/") == -1:
-                    resourcess.append(f"{a}/{e}")
-                else:
-                    resourcess.append(f"{e}")
+                if e.find(f"{a}/") == -1: resourcess.append(f"{a}/{e}")
+                else: resourcess.append(f"{e}")
             web_accessible_resources = web_accessible_resources + resourcess
-for a, v in ses.items():
-    v["bundleAssociatedName"] = a
+for a, v in ses.items(): v["bundleAssociatedName"] = a
 for a, v in mans.items():
     if v.get("permissions"):
         for i in v.get("permissions"):
-            if not (i in manifest_permissions):
-                manifest_permissions.append(i)
+            if not (i in manifest_permissions): manifest_permissions.append(i)
 se["extensions"] = extensions_added
 ma["content_scripts"] = [
     {
@@ -136,52 +113,41 @@ printMainMessage(f"Manifest Permissions: {json.dumps(manifest_permissions)}")
 printMainMessage(f"Content Scripts/CSS: {json.dumps(content_scripts)} + {json.dumps(content_css)}")
 printMainMessage(f"Web Accessible Resources: {json.dumps(web_accessible_resources)}")
 printMainMessage("Saving Settings JSON..")
-with open(os.path.join(extension_path, "settings.json"), "w") as f:
-    json.dump(se, f, indent=4)
+with open(os.path.join(extension_path, "settings.json"), "w") as f: json.dump(se, f, indent=4)
 printMainMessage("Saving Manifest JSON..")
-with open(os.path.join(extension_path, "manifest.json"), "w") as f:
-    json.dump(ma, f, indent=4)
+with open(os.path.join(extension_path, "manifest.json"), "w") as f: json.dump(ma, f, indent=4)
     
 # Configuring Mini Extensions
 printWarnMessage("--- Configuring Mini Extensions ---")
 ex_name_replace_key = "{extension_name_this_is_replace_when_building_bundle_with_folder_name_if_youre_wondering}"
 for i in extensions_added:
     printMainMessage(f"Configuring {i}..")
-    with open(f"{extension_path}/{i}/settings.js", "w") as f:
-        f.write("// This file was used for the mini extensions but since this extension is currently being used for bundle extension, it's cleared with this message.")
+    with open(f"{extension_path}/{i}/settings.js", "w") as f: f.write("// This file was used for the mini extensions but since this extension is currently being used for bundle extension, it's cleared with this message.")
+    with open(f"{extension_path}/{i}/main.js", "w") as f: f.write("// This file was used for the mini extensions but since this extension is currently being used for bundle extension, it's cleared with this message.")
     for q in os.listdir(f"{extension_path}/{i}/"):
         if os.path.isfile(f"{extension_path}/{i}/{q}") and q.endswith(".js"):
-            with open(f"{extension_path}/{i}/{q}", "r") as f:
-                fi_f = f.read()
+            with open(f"{extension_path}/{i}/{q}", "r") as f: fi_f = f.read()
             fi_f = fi_f.replace(ex_name_replace_key, i)
-            with open(f"{extension_path}/{i}/{q}", "w") as f:
-                f.write(fi_f)
+            with open(f"{extension_path}/{i}/{q}", "w") as f: f.write(fi_f)
 
 # Final Touches
 printWarnMessage("--- Making Final Touches ---")
 for i in extensions_added:
     printMainMessage(f"Saving JSONs for {i}..")
-    with open(f"{extension_path}/{i}/org_manifest.json", "w") as f:
-        json.dump(mans[i], f, indent=4)
-    with open(f"{extension_path}/{i}/settings.json", "w") as f:
-        json.dump(ses[i], f, indent=4)
+    with open(f"{extension_path}/{i}/org_manifest.json", "w") as f: json.dump(mans[i], f, indent=4)
+    with open(f"{extension_path}/{i}/settings.json", "w") as f: json.dump(ses[i], f, indent=4)
     os.remove(f"{extension_path}/{i}/manifest.json")
 
 if not (os.name == "nt"):
     printMainMessage("Creating Bundle ZIP File..")
-    if os.path.exists(
-        os.path.join(current_path_location, "zip", "chromeExtension.zip")
-    ):
-        os.remove(os.path.join(current_path_location, "zip", "chromeExtension.zip"))
+    if os.path.exists(os.path.join(current_path_location, "zip", "chromeExtension.zip")): os.remove(os.path.join(current_path_location, "zip", "chromeExtension.zip"))
     a = subprocess.run(
         f'zip -r ../zip/chromeExtension.zip ./ -x "*.git*" -x "*.DS_Store" -x "__MACOSX*" -x "__pycache*"',
         cwd=extension_path,
         shell=True,
     )
-    if a.returncode == 0: 
-        printMainMessage("Ziping has succeeded!")
-    else:
-        printErrorMessage(f"Uh oh! Ziping has failed due to an error! Status Code: {a.returncode}")
+    if a.returncode == 0:  printMainMessage("Zipping has succeeded!")
+    else: printErrorMessage(f"Uh oh! Zipping has failed due to an error! Status Code: {a.returncode}")
 
 # Done!
 printSuccessMessage("SUCCESS!")
